+++ src/index.ts
@@ -1,8 +1,7 @@
-import type {Schema} from './adapters';
+import type {Schema} from './registry';
 
-import {wrap} from './adapters';
-
-export type {Schema} from './adapters';
+import {wrap} from './wrap';
+export type {Schema} from './registry';
 export type Infer<TSchema> = TSchema extends Schema<infer T> ? T : never;
 
 export async function assert<T>(schema: Schema<T>, data: unknown): Promise<T> {
